+++ Abs/Entities/PlayerEntity.cpp
@@ -1,5 +1,5 @@
 #include "PlayerEntity.h"
 
-Abs::PlayerEntity::PlayerEntity(float x, float y, int w, int h) : Entity(x, y, w, h) {
+Abs::PlayerEntity::PlayerEntity(float x, float y, float w, float h) : Entity(x, y, w, h) {
 
 }
