+++ Abs/Entities/PlayerShip.cpp
@@ -1,5 +1,5 @@
 #include "PlayerShip.h"
 
-Abs::PlayerShip::PlayerShip(float x, float y, int w, int h) : PlayerEntity(x, y, w, h){
+Abs::PlayerShip::PlayerShip(float x, float y, float w, float h) : PlayerEntity(x, y, w, h){
 
 }
